Método de buscar por tutor e doador adicionado

+++ src/main/java/br/com/cidadeAcolhedora/CidadeAcolhedora/Animal/AnimalRepository.java
@@ -21,9 +21,9 @@ public interface AnimalRepository extends CrudRepository<Animal, Long>{
 	public Optional<List<Animal>> findByEstado(@Param("id_estado") int id_estado);
 	
 	@Query("FROM Animal WHERE id_doador = :id_doador")
-	public Optional<List<Animal>> findById_doador(@Param("id_doador") int id_doador);
+	public Optional<List<Animal>> findById_doador(@Param("id_doador") Long id_doador);
 	
 	@Query("FROM Animal WHERE id_tutor = :id_tutor")
-	public Optional<List<Animal>> findById_tutor(@Param("id_tutor") int id_tutor);
+	public Optional<List<Animal>> findById_tutor(@Param("id_tutor") Long id_tutor);
 	
 }
